Erased unused libraries
