+++ QuestUp/Assets/[My Assets]/Scenes/Transition.unity
@@ -272,11 +272,6 @@ MonoBehaviour:
   m_RequiresDepthTexture: 0
   m_RequiresColorTexture: 0
   m_Version: 2
---- !u!1 &1387005566 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 1513029137993506617, guid: aa9712655dda8584bb500e6aaf8b3b14, type: 3}
-  m_PrefabInstance: {fileID: 1513029138418478113}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1435177798
 GameObject:
   m_ObjectHideFlags: 0
@@ -286,10 +281,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1435177799}
-  - component: {fileID: 1435177800}
-  - component: {fileID: 1435177801}
   m_Layer: 0
-  m_Name: Transition Controller
+  m_Name: Transition Manager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -310,35 +303,154 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1435177800
+--- !u!1 &1779579591
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1779579592}
+  - component: {fileID: 1779579595}
+  - component: {fileID: 1779579594}
+  - component: {fileID: 1779579593}
+  m_Layer: 5
+  m_Name: Timer
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!224 &1779579592
+RectTransform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1779579591}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 475783091075279613}
+  m_RootOrder: 2
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: 772, y: 368}
+  m_SizeDelta: {x: 200, y: 50}
+  m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &1779579593
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1435177798}
+  m_GameObject: {fileID: 1779579591}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ae1f9bd5e91130c42a5e364a1d8b10b0, type: 3}
+  m_Script: {fileID: 11500000, guid: 68227d04e5d2a6940bca4a5eec52c234, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _levelSO: {fileID: 11400000, guid: d902b02bebcc51747abc0fe99f4d6dab, type: 2}
-  _loadSceneAdditive: {fileID: 0}
-  _transitionAnimations: {fileID: 1435177801}
-  _intermissionCanvas: {fileID: 1387005566}
---- !u!114 &1435177801
+--- !u!114 &1779579594
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1435177798}
+  m_GameObject: {fileID: 1779579591}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c9657d076c03b304aae06fbce4854896, type: 3}
+  m_Script: {fileID: 11500000, guid: f4688fdb7df04437aeb418b961361dc5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _animator: {fileID: 7191597246420523701}
+  m_Material: {fileID: 0}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
+  m_RaycastTarget: 1
+  m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
+  m_Maskable: 1
+  m_OnCullStateChanged:
+    m_PersistentCalls:
+      m_Calls: []
+  m_text: 5
+  m_isRightToLeft: 0
+  m_fontAsset: {fileID: 11400000, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
+  m_sharedMaterial: {fileID: 2180264, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
+  m_fontSharedMaterials: []
+  m_fontMaterial: {fileID: 0}
+  m_fontMaterials: []
+  m_fontColor32:
+    serializedVersion: 2
+    rgba: 4294967295
+  m_fontColor: {r: 1, g: 1, b: 1, a: 1}
+  m_enableVertexGradient: 0
+  m_colorMode: 3
+  m_fontColorGradient:
+    topLeft: {r: 1, g: 1, b: 1, a: 1}
+    topRight: {r: 1, g: 1, b: 1, a: 1}
+    bottomLeft: {r: 1, g: 1, b: 1, a: 1}
+    bottomRight: {r: 1, g: 1, b: 1, a: 1}
+  m_fontColorGradientPreset: {fileID: 0}
+  m_spriteAsset: {fileID: 0}
+  m_tintAllSprites: 0
+  m_StyleSheet: {fileID: 0}
+  m_TextStyleHashCode: -1183493901
+  m_overrideHtmlColors: 0
+  m_faceColor:
+    serializedVersion: 2
+    rgba: 4294967295
+  m_fontSize: 72
+  m_fontSizeBase: 72
+  m_fontWeight: 400
+  m_enableAutoSizing: 0
+  m_fontSizeMin: 18
+  m_fontSizeMax: 72
+  m_fontStyle: 0
+  m_HorizontalAlignment: 4
+  m_VerticalAlignment: 256
+  m_textAlignment: 65535
+  m_characterSpacing: 0
+  m_wordSpacing: 0
+  m_lineSpacing: 0
+  m_lineSpacingMax: 0
+  m_paragraphSpacing: 0
+  m_charWidthMaxAdj: 0
+  m_enableWordWrapping: 0
+  m_wordWrappingRatios: 0.4
+  m_overflowMode: 0
+  m_linkedTextComponent: {fileID: 0}
+  parentLinkedComponent: {fileID: 0}
+  m_enableKerning: 1
+  m_enableExtraPadding: 0
+  checkPaddingRequired: 0
+  m_isRichText: 1
+  m_parseCtrlCharacters: 1
+  m_isOrthographic: 1
+  m_isCullingEnabled: 0
+  m_horizontalMapping: 0
+  m_verticalMapping: 0
+  m_uvLineOffset: 0
+  m_geometrySortingOrder: 0
+  m_IsTextObjectScaleStatic: 0
+  m_VertexBufferAutoSizeReduction: 0
+  m_useMaxVisibleDescender: 1
+  m_pageToDisplay: 1
+  m_margin: {x: 0, y: 0, z: 0, w: 0}
+  m_isUsingLegacyAnimationComponent: 0
+  m_isVolumetricText: 0
+  m_hasFontAssetChanged: 0
+  m_baseMaterial: {fileID: 0}
+  m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
+--- !u!222 &1779579595
+CanvasRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1779579591}
+  m_CullTransparentMesh: 1
 --- !u!1 &1974180397
 GameObject:
   m_ObjectHideFlags: 0
@@ -408,6 +520,22 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 475783090073368182, guid: ec6fb6b6cde9621468b13eb8d5d952cf, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: -795
+      objectReference: {fileID: 0}
+    - target: {fileID: 475783090073368182, guid: ec6fb6b6cde9621468b13eb8d5d952cf, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 368
+      objectReference: {fileID: 0}
+    - target: {fileID: 475783090756050220, guid: ec6fb6b6cde9621468b13eb8d5d952cf, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: -795
+      objectReference: {fileID: 0}
+    - target: {fileID: 475783090756050220, guid: ec6fb6b6cde9621468b13eb8d5d952cf, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 269.99997
+      objectReference: {fileID: 0}
     - target: {fileID: 475783091574591331, guid: ec6fb6b6cde9621468b13eb8d5d952cf, type: 3}
       propertyPath: m_Pivot.x
       value: 0
@@ -498,6 +626,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: ec6fb6b6cde9621468b13eb8d5d952cf, type: 3}
+--- !u!224 &475783091075279613 stripped
+RectTransform:
+  m_CorrespondingSourceObject: {fileID: 475783091574591331, guid: ec6fb6b6cde9621468b13eb8d5d952cf, type: 3}
+  m_PrefabInstance: {fileID: 475783091075279612}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &1513029138418478113
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -724,8 +857,3 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 62f3e4ee2bc5f0d458e240ae2b241489, type: 3}
---- !u!95 &7191597246420523701 stripped
-Animator:
-  m_CorrespondingSourceObject: {fileID: 8952937063621192262, guid: 62f3e4ee2bc5f0d458e240ae2b241489, type: 3}
-  m_PrefabInstance: {fileID: 7191597246420523700}
-  m_PrefabAsset: {fileID: 0}
