+++ QuestUp/Assets/[My Assets]/[Scripts]/Managers/LevelManager.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using UnityEditor.ShaderGraph.Internal;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
@@ -8,6 +9,8 @@ namespace QuestUp
     public class LevelManager : MonoBehaviour
     {
         public static LevelManager Instance { get; private set; }
+
+        [field: SerializeField] public LevelSO CurrentLevelSO { get; private set; }
 
         private void Awake()
         {
@@ -19,23 +22,10 @@ namespace QuestUp
             DontDestroyOnLoad(gameObject);
         }
 
-        public void NextLevel(int previousLevel, int nextLevel)
+        public void StartLevelTimer()
         {
-            LoadLevel(nextLevel);
-            UnloadLevel(previousLevel);
-        }
-
-        private void LoadLevel(int nextLevel)
-        {
-            SceneManager.LoadSceneAsync(nextLevel, LoadSceneMode.Additive);
-        }
-
-        private void UnloadLevel(int previousLevel)
-        {
-            if (previousLevel != 0 && SceneManager.GetActiveScene() == SceneManager.GetSceneByBuildIndex(previousLevel))
-            {
-                SceneManager.UnloadSceneAsync(previousLevel);
-            }
+            float levelDuration = CurrentLevelSO.LevelDuration;
+            //Timer.OnTimerEnd(_levelDuration, () => StartCoroutine(GoToLevel(1)));
         }
     }
 }
