+++ QuestUp/Assets/[My Assets]/[Scripts]/Managers/TransitionManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ae1f9bd5e91130c42a5e364a1d8b10b0
+guid: 2ab94d16caea7ec47bf5f344c83877b7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
