+++ QuestUp/Assets/[My Assets]/[Scripts]/UI/TimerUI.cs
@@ -3,18 +3,13 @@ using TMPro;
 
 namespace QuestUp
 {
-    public class HealthUI : MonoBehaviour
+    public class TimerUI : MonoBehaviour
     {
         private TMP_Text _text = null;
 
         private void Start()
         {
             _text = GetComponent<TMP_Text>();
-        }
-
-        void Update()
-        {
-            _text.text = "Health: " + HealthManager.Instance.CurrentHealth;
         }
     }
 }
