+++ QuestUp/Assets/[My Assets]/[Scripts]/UI/TimerUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6734b7b71ea2f234695e9ba16eaab3c9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
